streamts: -fix 64bits (thx Houdini) - clean widgets

+++ src/driver/streamts.cpp
@@ -261,7 +261,7 @@ void streamts_main_thread(void * /*data*/)
 					tcnt++;
 					exit_flag = 0;
 					
-					pthread_create(&st, NULL, streamts_live_thread, (void *) connfd);
+					pthread_create(&st, NULL, streamts_live_thread, (void *) (long)connfd);
 				} 
 				else 
 				{
@@ -300,7 +300,7 @@ void * streamts_live_thread(void * data)
 	int pids[MAXPIDS];
 	char cbuf[512];
 	char *bp;
-	int fd = (int) data;
+	int fd = (int) (long)data;
 	FILE * fp;
 	unsigned char demuxfd_count = 0;
 
@@ -425,7 +425,7 @@ void streamts_file_thread(void *data)
 	int tsfilelen = 0;
 	int fileslice = 0;
 	int i = 0;
-	int fd = (int) data;
+	int fd = (int) (long)data;
 
 	buf = (unsigned char *) malloc(IN_SIZE);
